+++ mrv2/lib/mrvFlmm/Flmm_ColorA_Chooser.cxx
@@ -115,7 +115,7 @@ static Fl_Menu_Item mode_menu[] = {
 
 int Flmm_Value_Input::format(char* buf) {
   Flmm_ColorA_Chooser* c = (Flmm_ColorA_Chooser*)parent();
-  if (c->mode() == M_HEX) return sprintf(buf,"0x%02X", int(value()));
+  if (c->mode() == M_HEX) return snprintf(buf, 5, "0x%02X", int(value()));
   else return Fl_Valuator::format(buf);
 }